+++ packages/router/src/index.ts
@@ -4,4 +4,8 @@ export { RouterOptions };
 import createRouter from "./router/createRouter";
 import prepareRoutes from "./prepare/prepareRoutes";
 
-export { createRouter, prepareRoutes };
+import announce from "./side-effects/announce";
+import scroll from "./side-effects/scroll";
+import title from "./side-effects/title";
+
+export { createRouter, prepareRoutes, announce, scroll, title };