+++ packages/side-effects/side-effect-scroll/types/index.d.ts
@@ -1,2 +1,0 @@
-import { Observer } from "@curi/types";
-export default function createScrollSideEffect(): Observer;