+++ packages/side-effects/side-effect-title/types/index.d.ts
@@ -1,3 +1,0 @@
-import { Observer, Emitted } from "@curi/types";
-export declare type TitleBuilder = (emitted: Emitted) => string;
-export default function createTitleSideEffect(callback: TitleBuilder): Observer;