+++ packages/types/src/index.ts
@@ -46,6 +46,8 @@ export interface ResponseHandlerOptions {
 }
 // an observer function that will be called when there is a new navigation
 export type Observer = (props?: Emitted) => void;
+// a SideEffect is like an Observer, but will always be called with an Emitted
+export type SideEffect = (props: Emitted) => void;
 
 export interface ResponseAndNav {
   response: Response;