+++ website/src/components/links/collapsible/PackageLinks.js
@@ -22,7 +22,7 @@ const GroupPackages = ({ packages }) => (
 function PackageLinks() {
   const groups = PACKAGE_API.grouped();
   return Object.keys(groups)
-    .filter(t => t !== "route interactions")
+    .filter(t => t !== "route interactions" && t !== "side effects")
     .map(title => (
       <CollapsibleGroup key={title} title={title} initial={false}>
         <GroupPackages packages={groups[title]} />