+++ website/src/pages/Packages/router/v2/component.js
@@ -4,6 +4,9 @@ import { About, APIBlock, IJS } from "../../../../components/package/common";
 import { CreateRouterAPI } from "./api/createRouter";
 import { PrepareRoutesAPI } from "./api/prepareRoutes";
 import { RoutePropertiesAPI } from "./api/route-objects";
+import { AnnounceAPI } from "./api/announce";
+import { ScrollAPI } from "./api/scroll";
+import { TitleAPI } from "./api/title";
 
 export default class RouterPkg extends React.PureComponent {
   render() {
@@ -17,6 +20,9 @@ export default class RouterPkg extends React.PureComponent {
         <APIBlock>
           <PrepareRoutesAPI />
           <CreateRouterAPI />
+          <AnnounceAPI />
+          <ScrollAPI />
+          <TitleAPI />
           <RoutePropertiesAPI />
         </APIBlock>
       </React.Fragment>