+++ website/src/pages/Packages/router/v2/contents.js
@@ -1,6 +1,9 @@
 import { meta as createRouterMeta } from "./api/createRouter";
 import { meta as prepareRoutesMeta } from "./api/prepareRoutes";
 import { meta as RoutePropertiesMeta } from "./api/route-objects";
+import { meta as announceMeta } from "./api/announce";
+import { meta as scrollMeta } from "./api/scroll";
+import { meta as titleMeta } from "./api/title";
 
 export default [
   {
@@ -14,6 +17,13 @@ export default [
   {
     title: "API",
     hash: "API",
-    children: [prepareRoutesMeta, createRouterMeta, RoutePropertiesMeta]
+    children: [
+      prepareRoutesMeta,
+      createRouterMeta,
+      announceMeta,
+      scrollMeta,
+      titleMeta,
+      RoutePropertiesMeta
+    ]
   }
 ];