+++ website/src/pages/Packages/side-effect-title/v2/index.js
@@ -1,4 +1,0 @@
-import SideEffectTitlePkg from "./component";
-import contents from "./contents";
-
-export { SideEffectTitlePkg as component, contents };