+++ website/src/router.js
@@ -1,21 +1,14 @@
 import { browser } from "@hickory/browser";
-import { createRouter } from "@curi/router";
-import titleSideEffect from "@curi/side-effect-title";
-import scrollSideEffect from "@curi/side-effect-scroll";
-import ariaLiveSideEffect from "@curi/side-effect-aria-live";
+import { createRouter, title, scroll, announce } from "@curi/router";
 
 import routes from "./routes";
 
-const setTitle = titleSideEffect(
-  ({ response }) => `${response.meta.title} | Curi Documentation`
-);
-const scrollTo = scrollSideEffect();
-const announce = ariaLiveSideEffect(
-  ({ response }) => `Navigated to ${response.meta.title}`
-);
-
 const router = createRouter(browser, routes, {
-  sideEffects: [setTitle, scrollTo, announce],
+  sideEffects: [
+    title(({ response }) => `${response.meta.title} | Curi Documentation`),
+    scroll(),
+    announce(({ response }) => `Navigated to ${response.meta.title}`)
+  ],
   invisibleRedirects: true
 });