+++ packages/router/CHANGELOG.md
@@ -1,3 +1,7 @@
+## Next
+
+* Merge the `aria-live` (now `announce`), `scroll`, and `title` side effects into package.
+
 ## 2.0.0-beta.19
 
 * `prepareRoutes` takes an array of routes as its first (and only) argument.